+++ BusinessLogicTier/Logic/Management/IManagementLogic.cs
@@ -179,6 +179,11 @@ public interface IManagementLogic
     void RemoveProjectClusterNodeTypeAggregation(long projectId, long clusterNodeTypeAggregationId);
     List<AccountingState> ListAccountingStates(long projectId);
 
+    string BackupDatabase();
+    string BackupDatabaseTransactionLogs();
+    List<DatabaseBackup> ListDatabaseBackups(DateTime? fromDateTime, DateTime? toDateTime, DatabaseBackupType type);
+    void RestoreDatabase(string backupFileName, bool includeLogs);
+
     Task<Status> Status(long projectId, DateTime? timeFrom, DateTime? timeTo);
 
     StatusCheckLogs StatusErrorLogs(long projectId, DateTime? timeFrom, DateTime? timeTo);