+++ DomainObjects/Management/DatabaseBackupType.cs
@@ -1,0 +1,7 @@
+﻿namespace HEAppE.DomainObjects.Management;
+
+public enum DatabaseBackupType
+{
+    Full,
+    Log
+}