+++ ExtModels/Management/Models/DatabaseBackupTypeExt.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel;
+
+namespace HEAppE.ExtModels.Management.Models;
+
+/// <summary>
+/// Database backup type
+/// </summary>
+[Description("Database backup type")]
+public enum DatabaseBackupTypeExt
+{
+    Full,
+    Log
+}