+++ RestApi/InputValidator/ManagementValidator.cs
@@ -77,6 +77,8 @@ public class ManagementValidator : AbstractValidator
             RemoveProjectClusterNodeTypeAggregationModel ext =>
                 ValidateRemoveProjectClusterNodeTypeAggregationModel(ext),
             AccountingStateModel ext => ValidateAccountingStateModel(ext),
+            ListDatabaseBackupsModel ext => ValidateListDatabaseBackupsModel(ext),
+            RestoreDatabaseModel ext => ValidateRestoreDatabaseModel(ext),
             ModifyClusterAuthenticationCredentialModel ext => ValidateModifyClusterAuthenticationCredentialModel(ext),
             StatusModel ext => ValidateStatusModel(ext),
             _ => string.Empty
@@ -769,6 +771,31 @@ public class ManagementValidator : AbstractValidator
         if (!sessionCodeValidation.IsValid) _messageBuilder.AppendLine(sessionCodeValidation.Message);
 
         ValidateId(model.ProjectId, "ProjectId");
+
+        return _messageBuilder.ToString();
+    }
+
+    private string ValidateListDatabaseBackupsModel(ListDatabaseBackupsModel model)
+    {
+        var sessionCodeValidation = new SessionCodeValidator(model.SessionCode).Validate();
+        if (!sessionCodeValidation.IsValid) _messageBuilder.AppendLine(sessionCodeValidation.Message);
+
+        if (model.ToDateTime.HasValue && model.ToDateTime < model.FromDateTime)
+            _messageBuilder.AppendLine($"{model.ToDateTime} must be greater than {model.FromDateTime}");
+
+        return _messageBuilder.ToString();
+    }
+
+    private string ValidateRestoreDatabaseModel(RestoreDatabaseModel model)
+    {
+        var sessionCodeValidation = new SessionCodeValidator(model.SessionCode).Validate();
+        if (!sessionCodeValidation.IsValid) _messageBuilder.AppendLine(sessionCodeValidation.Message);
+
+        if (string.IsNullOrEmpty(model.BackupFileName))
+            _messageBuilder.AppendLine($"BackupFileName can not be null or empty.");
+
+        if(ContainsIllegalCharactersForFileName(model.BackupFileName))
+            _messageBuilder.AppendLine($"BackupFileName contains illegal characters. Provide only file name.");
 
         return _messageBuilder.ToString();
     }